feat: iteration 1 - direct message, collect topics
"I have this Slack bot written in Python using their Bolt
framework. Add a feature so that I can use a slash command in Slack such
as "/forum start" and this triggers the bot to start a configurable
period of "topic gathering". During this period, the bot should listen
for direct messages from other users, and persist them to some store
including information about the current time, the message contents, and
its sender"