feat: iteration 5 - topic voting
"Now, instead of just announcing the topics when the period ends, a new
type of period begins where users should be able to cast votes for
topics they are interested in. A user should be able to vote for
multiple topics, regardless of who submitted them.

This voting period should also expire automatically after a configured
time, specified through a new argument to the "/forum start" command.
When it ends, the bot should announce a list of the winning topics,
sorted in descending order by the number of votes each item had."