feat: iteration 6 - calendar invites
"I want to add functionality so that when the voting period ends, the
application uses Microsoft Graph API to create Outlook Calendar events
for the top N winning topics, and the Slack users who voted for each
topic are invited as participants to the corresponding event.

The date and time for a topic's event should be provided from a pool of
slots. The pool should be configurable so that I can specify a list of
entries with fields such as "day_of_week", "start_time", "end_time". The
slot information can be loaded from a JSON file. We can assume that the
Slack users have the same e-mail in Outlook and on Slack.

After a calendar event has been successfully created, the application
should announce a summary of it on Slack."